Añadido:Restauracion y elimincacion permanente de proyectos

+++ app/Http/Controllers/ProjectController.php
@@ -21,7 +21,8 @@ class ProjectController extends Controller
     {
         return view('projects.index', [
             'newProject' => new Project(),
-            'projects' => Project::with('category')->latest()->paginate()
+            'projects' => Project::with('category')->latest()->paginate(),
+            'deletedProjects' => Project::onlyTrashed()->get()
         ]);
     }
 
@@ -34,7 +35,7 @@ class ProjectController extends Controller
 
         $project = new Project($request->validated());
 
-        $this -> authorize('create', $project);
+        $this->authorize('create', $project);
 
         if ($request->hasFile('image')) {
             $project->image = $request->file('image')->store('images', 'public');
@@ -80,21 +81,40 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        $this -> authorize('create', $project);
-
-
-        if ($project->image) {
-            Storage::disk('public')->delete($project->image);
-        }
+        $this->authorize('create', $project);
 
         $project->delete();
 
         return redirect()->route('projects.index')->with('status', 'Proyecto eliminado');
     }
+    
+    public function restore($id)
+    {
+        $project = Project::withTrashed()->findOrFail($id);
+        $this->authorize('restore', $project);
+        $project->restore();
+    
+        return redirect()->route('projects.index')->with('status', 'Proyecto restaurado');
+    }
+    
+    public function forceDelete($id)
+    {
+        $project = Project::withTrashed()->findOrFail($id);
+        $this->authorize('forceDelete', $project);
+    
+        if ($project->image) {
+            Storage::disk('public')->delete($project->image);
+        }
+    
+        $project->forceDelete();
+    
+        return redirect()->route('projects.index')->with('status', 'Proyecto eliminado permanentemente');
+    }
+
 
     public function create()
     {
-        $this->authorize('create', $project = new Project()); 
+        $this->authorize('create', $project = new Project());
 
         return view('projects.create', [
             'project' => $project,
@@ -104,7 +124,7 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
-        $this -> authorize('update', $project);
+        $this->authorize('update', $project);
 
 
         return view('projects.edit', [
